app/grpc_server/requirements_grpc.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4